Aigle/0.2/AiModelLifecycle/requirements.txt
# ========================================
# AI Model Lifecycle Management Platform
# Python Dependencies
# ========================================

# ========================================
# Core Web Framework & API
# ========================================
fastapi
uvicorn
pydantic
python-multipart
# ========================================
# MLOps & Model Management
# ========================================
mlflow
ollama
lakefs-client
lakefs-spec

# ========================================
# Cloud Storage & Database
# ========================================
boto3 # mlflow s3 artifact store 需要
psycopg2-binary # mlflow postgres backend store 需要


# ========================================
# Machine Learning & Deep Learning
# ========================================
transformers
torch
torchvision
torchaudio
sentencepiece
accelerate
# datasets 和 pyarrow 版本兼容性要求
# datasets >= 2.19.0 支持 pyarrow 14-17
# 或者使用較舊但穩定的版本組合
pyarrow>=14.0.0,<18.0.0
datasets>=2.19.0

# ========================================
# Audio & Video Processing  
# ========================================
librosa
opencv-python
pillow
soundfile

# ========================================
# Configuration & Environment
# ========================================
pyyaml
python-dotenv
# ========================================
# HTTP Client & Networking
# ========================================
httpx
aiohttp
requests

# ========================================
# Data Processing & Utilities
# ========================================
numpy
pandas
scipy
scikit-learn

# ========================================
# Async & Concurrency
# ========================================
asyncio
aiofiles
celery

# ========================================
# Performance & Optimization
# ========================================
psutil